backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.11
django-ckeditor-5==0.2.13
Pillow==10.1.0
redis==5.0.1
hiredis==2.3.2
openai==1.12.0
httpx>=0.25.0,<0.28.0
django-auditlog==2.3.0
django-redis==5.4.0

# Data processing dependencies
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# GNN / Similar Companies dependencies
networkx>=3.0
node2vec>=0.4.6
scikit-learn>=1.3.0
lightgbm>=4.0.0
hdbscan>=0.8.33
tqdm>=4.65.0

# Testing dependencies
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.1.1
factory-boy==3.3.3
faker==33.3.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
